make clojure:swank available; use clojure 1.3.0
I saw perf_base64.clj go in under test, though it's not clear to me
how it's intended to be used since it's not actually called from any
unit tests. I decided that what I needed to do was be able to try it
out at the REPL while reading the source. To this end, I decided to
use swank/slime (as per my usual approach). mvn clojure:repl didn't do
what I required since test code isn't availble on the classpath,
therefore it can't be loaded and explored. ...

Upgrade parent to 0.0.25 so that the dependency is to clojure
1.3.0 (release) and not some earlier pre-release build.

Add a dependency to swank-clojure 1.3.3, since that appears to be the
newest release version.

Override clojure-maven-plugin to 1.3.8 since 1.3.7 configured by
parent 0.0.25 is incompatible with swank-clojure 1.3.3.

In fact, swank-clojure 1.3.1, 1.3.2, 1.3.3 all require
clojure-maven-plugin 1.3.8. Conversely swank-clojure 1.3.0 requires
clojure-maven-plugin 1.3.7.

    Exception in thread "main" java.lang.IllegalArgumentException: \
    No value supplied for key: true
    at clojure.lang.PersistentHashMap.createWithCheck(PersistentHashMap.java:89)

                      clojure-maven-plugin
    swank-clojure     1.3.7          1.3.8
                      -----          -----
            1.3.0 |   works          fails
            1.3.1 |   fails          works
            1.3.2 |   fails          works
            1.3.3 |   fails          works

Incidentally, one of my reasons for choosing to try swank-clojure
1.3.3 was the hope that it might work cause the doc macro to work
again from the repl. This has been broken since upgrading to Clojure
1.3.0. Alas no, though manually executing (require 'clojure.repl) at
the repl prompt in every namespace "fixes" the problem.